+++ frontend/app/components/Dashboard/components/FilterSeries/FilterSeries.tsx
@@ -4,12 +4,12 @@ import { observer } from 'mobx-react-lite';
 import { Button, Divider, Space, Card } from 'antd';
 import { ChevronDown, ChevronUp, Trash } from 'lucide-react';
 import SeriesName from './SeriesName';
-import FilterListHeader from 'Shared/Filters/FilterList/FilterListHeader';
-import FilterSelection from 'Shared/Filters/FilterSelection';
-import { Filter } from '@/mstore/types/filterConstants';
-import { Plus } from 'lucide-react';
-import UnifiedFilterList from 'Shared/Filters/FilterList/UnifiedFilterList';
 import { useStore } from '@/mstore';
+import { Plus } from 'lucide-react';
+import { Filter } from '@/mstore/types/filterConstants';
+import FilterSelection from 'Shared/Filters/FilterSelection';
+import FilterListHeader from 'Shared/Filters/FilterList/FilterListHeader';
+import UnifiedFilterList from 'Shared/Filters/FilterList/UnifiedFilterList';
 import IFilterSeries from '@/mstore/types/filterSeries';
 import FilterItem from '@/mstore/types/filterItem';