+++ frontend/app/components/Overview/Overview.tsx
@@ -1,20 +1,13 @@
 import React from 'react';
 import withPageTitle from 'HOCs/withPageTitle';
 import SessionsTabOverview from 'Shared/SessionsTabOverview/SessionsTabOverview';
-import FFlagsList from 'Components/FFlags';
-import NewFFlag from 'Components/FFlags/NewFFlag';
 import { Switch, Route } from 'react-router';
 import {
   sessions,
-  fflags,
   withSiteId,
-  newFFlag,
-  fflag,
-  fflagRead,
   bookmarks
 } from 'App/routes';
 import { withRouter, RouteComponentProps, useLocation } from 'react-router-dom';
-import FlagView from 'Components/FFlags/FlagView/FlagView';
 import { observer } from 'mobx-react-lite';
 import { useStore } from '@/mstore';
 import Bookmarks from 'Shared/SessionsTabOverview/components/Bookmarks/Bookmarks';
@@ -25,22 +18,19 @@ interface IProps extends RouteComponentProps {
   match: {
     params: {
       siteId: string;
-      fflagId?: string;
     };
   };
 }
 
 // TODO should move these routes to the Routes file
 function Overview({ match: { params } }: IProps) {
-  const { searchStore, filterStore, projectsStore } = useStore();
-  const { siteId, fflagId } = params;
+  const { searchStore, projectsStore } = useStore();
+  const { siteId } = params;
   const location = useLocation();
   const tab = location.pathname.split('/')[2];
-  const { activeSiteId } = projectsStore;
 
   React.useEffect(() => {
     searchStore.setActiveTab(tab);
-    // void filterStore.fetchFilters(activeSiteId + '');
   }, [tab]);
 
   return (
@@ -54,18 +44,6 @@ function Overview({ match: { params } }: IProps) {
         <div className="mb-5 w-full mx-auto" style={{ maxWidth: PANEL_SIZES.maxWidth }}>
           <Bookmarks />
         </div>
-      </Route>
-      <Route exact strict path={withSiteId(fflags(), siteId)}>
-        <FFlagsList siteId={siteId} />
-      </Route>
-      <Route exact strict path={withSiteId(newFFlag(), siteId)}>
-        <NewFFlag siteId={siteId} />
-      </Route>
-      <Route exact strict path={withSiteId(fflag(), siteId)}>
-        <NewFFlag siteId={siteId} fflagId={fflagId} />
-      </Route>
-      <Route exact strict path={withSiteId(fflagRead(), siteId)}>
-        <FlagView siteId={siteId} fflagId={fflagId!} />
       </Route>
     </Switch>
   );